Tests: Move legacy unveil-symlinks test to TestKernelUnveil
Move the legacy kernel test unveil-symlinks into a new case in
TestKernelUnveil called test_symlinks.

The original test called `unveil(nullptr, nullptr)` to disable future
unveil calls, but this test suite has other calls to unveil that may
break depending on the execution order of the test cases, so that is
omitted. There can be a separate test for that functionality, since
it's unrelated to symlinks anyway.